requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
Django==4.2.2
django-filter==23.2
django-graphql-jwt==0.3.4
graphene==3.2.2
graphene-django==3.1.2
graphql-core==3.2.3
graphql-relay==3.2.0
promise==2.3
PyJWT==2.7.0
python-decouple==3.8
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.6.3
